requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
supabase==2.0.2
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.24.1
sentence-transformers==2.2.2
pypdf==3.17.4
numpy==1.24.3
python-dotenv==1.0.0
torch==2.7.0
torchaudio==2.7.0
huggingface-hub==0.16.4